+++ Pyomic/bulk/__init__.py
@@ -5,9 +5,9 @@ bulk (A omic framework for bulk omic analysis)
 #from Pyomic.bulk.DeGene import find_DEG,Density_norm,Plot_gene_expression,ID_mapping
 #from Pyomic.bulk.Gene_module import pywgcna
 
-from ._Gene_module import pywgcna
-from ._Enrichment import geneset_enrichment,geneset_plot
-from ._network import string_interaction,string_map,generate_G
+from ._Gene_module import pyWGCNA
+from ._Enrichment import pyGSEA,geneset_enrichment,geneset_plot
+from ._network import pyPPI,string_interaction,string_map,generate_G
 from ._chm13 import get_chm13_gene,find_chm13_gene
-from ._Deseq2 import pyDEseq,deseq2_normalize,estimateSizeFactors,estimateDispersions,Matrix_ID_mapping
-from ._tcga import TCGA
+from ._Deseq2 import pyDEG,deseq2_normalize,estimateSizeFactors,estimateDispersions,Matrix_ID_mapping
+from ._tcga import pyTCGA