+++ Pyomic/bulk2single/_bulk2single.py
@@ -8,6 +8,8 @@ from ._vae import train_vae, generate_vae, load_vae
 #from .map_utils import create_data, DFRunner, joint_analysis, knn
 import os
 import warnings
+import matplotlib
+from typing import Union,Tuple
 warnings.filterwarnings("ignore")
 
 
@@ -16,29 +18,19 @@ class Bulk2Single:
     Bulk2Single class.
     
     """
-    def __init__(self,bulk_data,single_data,celltype_key,
-                 top_marker_num=500,ratio_num=1,gpu=0):
+    def __init__(self,bulk_data:pd.DataFrame,single_data:anndata.AnnData,celltype_key:str,
+                 top_marker_num:int=500,ratio_num:int=1,gpu:Union[int,str]=0):
         """
         Initializes the Bulk2Single class.
 
-        Parameters
-        ----------
-        - bulk_data: `pandas.DataFrame`
-            The bulk RNA-seq data.
-        - single_data: `pandas.DataFrame`
-            The single-cell RNA-seq data.
-        - celltype_key: `str`
-            The name of the column in the bulk data containing cell types.
-        - top_marker_num: `int`, optional
-            The number of top markers to select per cell type. Default is 500.
-        - ratio_num: `float`, optional
-            The ratio between the number of single cells and target number of converted cells. Default is 1.
-        - gpu: `int/str`, optional
-            The ID of the GPU to use. Set to -1 to use CPU. Default is 0. If set to 'mps', the MPS backend will be used.
+        Arguments:
+            bulk_data: The bulk RNA-seq data.
+            single_data: The single-cell RNA-seq data.
+            celltype_key: The name of the column in the bulk data containing cell types.
+            top_marker_num: The number of top markers to select per cell type. Default is 500.
+            ratio_num: The ratio between the number of single cells and target number of converted cells. Default is 1.
+            gpu: The ID of the GPU to use. Set to -1 to use CPU. Default is 0. If set to 'mps', the MPS backend will be used.
 
-        Returns
-        -------
-        - None
         """
         self.bulk_data=bulk_data
         self.single_data=single_data
@@ -53,43 +45,32 @@ class Bulk2Single:
         self.history=[]
 
     def train(self,
-            vae_save_dir='save_model',
-            vae_save_name='vae',
-            generate_save_dir='output',
-            generate_save_name='output',
-            batch_size=512,
-            learning_rate=1e-4,
-            hidden_size=256,
-            epoch_num=5000,
-            patience=50):
+            vae_save_dir:str='save_model',
+            vae_save_name:str='vae',
+            generate_save_dir:str='output',
+            generate_save_name:str='output',
+            batch_size:int=512,
+            learning_rate:int=1e-4,
+            hidden_size:int=256,
+            epoch_num:int=5000,
+            patience:int=50,save:bool=True)->torch.nn.Module:
         """
         Trains the VAE model.
 
-        Parameters
-        ----------
-        - vae_save_dir: `str`, optional
-            The directory to save the trained VAE model. Default is 'save_model'.
-        - vae_save_name: `str`, optional
-            The name of the saved VAE model. Default is 'vae'.
-        - generate_save_dir: `str`, optional
-            The directory to save the generated single-cell data. Default is 'output'.
-        - generate_save_name: `str`, optional
-            The name of the saved generated single-cell data. Default is 'output'.
-        - batch_size: `int`, optional
-            The batch size for training. Default is 512.
-        - learning_rate: `float`, optional
-            The learning rate for training. Default is 1e-4.
-        - hidden_size: `int`, optional
-            The hidden size for the encoder and decoder networks. Default is 256.
-        - epoch_num: `int`, optional
-            The maximum number of epochs for training. Default is 5000.
-        - patience: `int`, optional
-            The number of epochs to wait before early stopping. Default is 50.
+        Arguments:
+            vae_save_dir: The directory to save the trained VAE model. Default is 'save_model'.
+            vae_save_name: The name of the saved VAE model. Default is 'vae'.
+            generate_save_dir: The directory to save the generated single-cell data. Default is 'output'.
+            generate_save_name: The name of the saved generated single-cell data. Default is 'output'.
+            batch_size: The batch size for training. Default is 512.
+            learning_rate: The learning rate for training. Default is 1e-4.
+            hidden_size: The hidden size for the encoder and decoder networks. Default is 256.
+            epoch_num: The maximum number of epochs for training. Default is 5.
+            patience: The number of epochs to wait before early stopping. Default is 50.
+            save: Whether to save the trained VAE model. Default is True.
 
-        Returns
-        -------
-        - vae_net: `torch.nn.Module`
-            The trained VAE model.
+        Returns:
+            vae_net: The trained VAE model.
         """
         single_cell, label, breed_2_list, index_2_gene, cell_number_target_num, \
         nclass, ntrain, feature_size = self.__get_model_input(self.input_data, self.cell_target_num)
@@ -104,11 +85,12 @@ class Bulk2Single:
                             hidden_size=hidden_size,
                             patience=patience,)
         print('...vae training done!')
-        path_save = os.path.join(vae_save_dir, f"{vae_save_name}.pth")
-        if not os.path.exists(vae_save_dir):
-            os.makedirs(vae_save_dir)
-        torch.save(vae_net.state_dict(), path_save)
-        print(f"...save trained vae in {path_save}.")
+        if save:
+            path_save = os.path.join(vae_save_dir, f"{vae_save_name}.pth")
+            if not os.path.exists(vae_save_dir):
+                os.makedirs(vae_save_dir)
+            torch.save(vae_net.state_dict(), path_save)
+            print(f"...save trained vae in {path_save}.")
         self.vae_net=vae_net
         self.history=history
         return vae_net
@@ -123,17 +105,28 @@ class Bulk2Single:
                                generate_save_dir, generate_save_name)
         return sc_g
     
-    def generate(self):
+    def save(self,vae_save_dir:str='save_model',
+            vae_save_name:str='vae',):
+        """
+        Saves the trained VAE model.
+
+        Arguments:
+            vae_save_dir: the directory to save the trained VAE model. Default is 'save_model'.
+            vae_save_name: the name of the saved VAE model. Default is 'vae'.
+
+        """
+        path_save = os.path.join(vae_save_dir, f"{vae_save_name}.pth")
+        if not os.path.exists(vae_save_dir):
+            os.makedirs(vae_save_dir)
+        torch.save(self.vae_net.state_dict(), path_save)
+        print(f"...save trained vae in {path_save}.")
+    
+    def generate(self)->anndata.AnnData:
         r"""
         Generate the single-cell data.
 
-        Parameters
-        ----------
-
-        Returns
-        -------
-        - sc_g: `anndata.AnnData`
-            The generated single-cell data.
+        Returns:
+            sc_g: The generated single-cell data.
         """
         single_cell, label, breed_2_list, index_2_gene, cell_number_target_num, \
         nclass, ntrain, feature_size = self.__get_model_input(self.input_data, self.cell_target_num)
@@ -146,21 +139,13 @@ class Bulk2Single:
         sc_g.obs[self.celltype_key] = generate_sc_meta.loc[sc_g.obs.index,'Cell_type'].values
         return sc_g
     
-    def load(self,vae_load_dir,hidden_size=256):
+    def load(self,vae_load_dir:str,hidden_size:int=256):
         r"""
         load the trained VAE model of Bulk2Single.
 
-        Parameters
-        ----------
-        - vae_load_dir: `str`
-            The directory to load the trained VAE model.
-        - hidden_size: `int`, optional
-            The hidden size for the encoder and decoder networks. Default is 256.
-
-        Returns
-        -------
-        - None
-
+        Arguments:
+            vae_load_dir: The directory to load the trained VAE model.
+            hidden_size: The hidden size for the encoder and decoder networks. Default is 256.
         """
         single_cell, label, breed_2_list, index_2_gene, cell_number_target_num, \
         nclass, ntrain, feature_size = self.__get_model_input(self.input_data, self.cell_target_num)
@@ -169,22 +154,17 @@ class Bulk2Single:
         self.vae_net=vae_net
 
     def load_and_generate(self,
-                              vae_load_dir,  # load_dir
-                              hidden_size=256):
+                              vae_load_dir:str,  # load_dir
+                              hidden_size:int=256)->anndata.AnnData:
         r"""
         load the trained VAE model of Bulk2Single and generate the single-cell data.
 
-        Parameters
-        ----------
-        - vae_load_dir: `str`
-            The directory to load the trained VAE model.
-        - hidden_size: `int`, optional
-            The hidden size for the encoder and decoder networks. Default is 256.
-        
-        Returns
-        -------
-        - sc_g: `anndata.AnnData`
-            The generated single-cell data.
+        Arguments:
+            vae_load_dir: The directory to load the trained VAE model.
+            hidden_size: The hidden size for the encoder and decoder networks. Default is 256.
+
+        Returns:
+            sc_g: The generated single-cell data.
         """
         single_cell, label, breed_2_list, index_2_gene, cell_number_target_num, \
         nclass, ntrain, feature_size = self.__get_model_input(self.input_data, self.cell_target_num)
@@ -202,26 +182,23 @@ class Bulk2Single:
         print('...generating done!')
         return sc_g
     
-    def plot_loss(self,figsize=(4,4)):
+    def plot_loss(self,figsize:tuple=(4,4))->Tuple[matplotlib.figure.Figure,matplotlib.axes._axes.Axes]:
         r"""
         plot the loss curve of the trained VAE model.
 
-        Parameters
-        ----------
-        - figsize: `tuple`, optional
-            The size of the figure. Default is (4,4).
-        
-        Returns
-        -------
-        - ax: `matplotlib.axes._subplots.AxesSubplot`
-            The axes of the figure.
+        Arguments:
+            figsize: The size of the figure. Default is (4,4).
+
+        Returns:
+            fig: The figure of the loss curve.
+            ax: The axes of the figure.
         """
         fig, ax = plt.subplots(figsize=figsize)
         ax.plot(range(len(self.history)),self.history)
         ax.set_title('Beta-VAE')
         ax.set_ylabel('Loss')
         ax.set_xlabel('Epochs')
-        return ax
+        return fig,ax
 
 
     def __get_model_input(self, data, cell_target_num):